[MOV] core, web: database manager functions next to the controllers
Some functions only exist in the database RPC service for the database
manager. Unexpose those function from the RPC service and move them next
to the database manager controller.

Remove the /web/session/get_lang_list controller as well, it is not used
anymore.
