+++ odoo/service/db.py
@@ -25,8 +25,6 @@ import odoo.release
 import odoo.sql_db
 import odoo.tools
 from odoo.exceptions import AccessDenied
-from odoo.release import version_info
-from odoo.sql_db import db_connect
 from odoo.tools import SQL, osutil
 from odoo.tools.date_utils import all_timezones
 from odoo.tools.misc import exec_pg_environ, find_pg_tool
@@ -482,50 +480,12 @@ def list_dbs(force=False):
             _logger.exception('Listing databases failed:')
             return []
 
-def list_db_incompatible(databases):
-    """"Check a list of databases if they are compatible with this version of Odoo
-
-        :param databases: A list of existing Postgresql databases
-        :return: A list of databases that are incompatible
-    """
-    incompatible_databases = []
-    server_version = '.'.join(str(v) for v in version_info[:2])
-    for database_name in databases:
-        with closing(db_connect(database_name).cursor()) as cr:
-            if odoo.tools.sql.table_exists(cr, 'ir_module_module'):
-                cr.execute("SELECT latest_version FROM ir_module_module WHERE name=%s", ('base',))
-                base_version = cr.fetchone()
-                if not base_version or not base_version[0]:
-                    incompatible_databases.append(database_name)
-                else:
-                    # e.g. 10.saas~15
-                    local_version = '.'.join(base_version[0].split('.')[:2])
-                    if local_version != server_version:
-                        incompatible_databases.append(database_name)
-            else:
-                incompatible_databases.append(database_name)
-    for database_name in incompatible_databases:
-        # release connection
-        odoo.sql_db.close_db(database_name)
-    return incompatible_databases
-
 
 def exp_list(document=False):
     if not odoo.tools.config['list_db']:
         raise odoo.exceptions.AccessDenied()
     return list_dbs()
 
-def exp_list_lang():
-    return odoo.tools.misc.scan_languages()
-
-def exp_list_countries():
-    list_countries = []
-    root = ET.parse(os.path.join(odoo.tools.config.root_path, 'addons/base/data/res_country_data.xml')).getroot()
-    for country in root.find('data').findall('record[@model="res.country"]'):
-        name = country.find('field[@name="name"]').text
-        code = country.find('field[@name="code"]').text
-        list_countries.append([code, name])
-    return sorted(list_countries, key=lambda c: c[1])
 
 def exp_server_version():
     """ Return the version of the server
@@ -540,7 +500,7 @@ def exp_server_version():
 def dispatch(method, params):
     g = globals()
     exp_method_name = 'exp_' + method
-    if method in ['db_exist', 'list', 'list_lang', 'server_version']:
+    if method in ['db_exist', 'list', 'server_version']:
         return g[exp_method_name](*params)
     elif exp_method_name in g:
         passwd = params[0]
